requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
academictorrents=2.3.3=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=0.4.3=pypi_0
altair=4.2.2=pypi_0
antlr4-python3-runtime=4.8=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
bencode-py=2.0.0=pypi_0
bitstring=3.1.5=pypi_0
blas=1.0=mkl
blinker=1.5=pypi_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38haa95532_0
charset-normalizer=3.0.1=pypi_0
clip=1.0=dev_0
colorama=0.4.6=py38haa95532_0
config=0.5.1=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
einops=0.3.0=pypi_0
entrypoints=0.4=pypi_0
filelock=3.9.0=pypi_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.16.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.16.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.1=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.2=pypi_0
imgaug=0.2.6=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
latent-diffusion=0.0.1=dev_0
libffi=3.4.2=hd77b12b_6
markdown=3.4.1=pypi_0
markdown-it-py=2.1.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=6.0.4=pypi_0
networkx=3.0=pypi_0
numpy=1.24.2=pypi_0
numpy-base=1.23.5=py38h4da318b_0
oauthlib=3.2.2=pypi_0
omegaconf=2.1.1=pypi_0
opencv-python=4.1.2.30=pypi_0
opencv-python-headless=4.7.0.68=pypi_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py38haa95532_0
protobuf=3.20.3=pypi_0
pudb=2019.2=pypi_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeck=0.8.0=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.14.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=pypi_0
pypubsub=3.3.0=pypi_0
python=3.8.16=h6244533_2
python-dateutil=2.8.2=pypi_0
pytorch-lightning=1.5.0=pypi_0
pytz=2022.7.1=pypi_0
pywavelets=1.4.1=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.3.1=pypi_0
rsa=4.9=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.10.0=pypi_0
semver=2.13.0=pypi_0
setuptools=65.6.3=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.40.1=h2bbff1b_0
streamlit=1.18.1=pypi_0
taming-transformers=0.0.1=dev_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
test-tube=0.7.5=pypi_0
tifffile=2023.2.3=pypi_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torch=1.13.1+cu116=pypi_0
torch-fidelity=0.3.0=pypi_0
torchaudio=0.13.1+cu116=pypi_0
torchmetrics=0.5.0=pypi_0
torchvision=0.14.1+cu116=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=py38haa95532_0
transformers=4.3.1=pypi_0
typing-extensions=4.5.0=pypi_0
urllib3=1.26.14=pypi_0
urwid=2.1.2=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py38haa95532_0
wincertstore=0.2=py38haa95532_2
yarl=1.8.2=pypi_0
zipp=3.14.0=pypi_0
